Updates boxes submodule (additional script)
The start.sh assumes the availability of the additional.sh script in the folder
of the Arch Linux/LXDE box, however, such a file did not yet exist. Now that it
does, the commits for that submodule needed to be applied before continuing the
testing of the start/clean scripts.